6382 | 7 files | M DumpSource2/schemas/client/CDOTA_Modifier_Bloodseeker_Sanguivore.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Bloodseeker_Thirst.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Bloodseeker_Sanguivore.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Bloodseeker_Thirst.h,  M game/bin/built_from_cl.txt,  M game/dota/pak01_dir.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/17947852/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Bloodseeker_Thirst.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Bloodseeker_Thirst
 {
-	int32 half_bonus_aoe;
+	float32 half_bonus_aoe;
 	int32 visibility_threshold_pct;
 	int32 invis_threshold_pct;
 	int32 min_bonus_pct;
